- Refactor from component to template-based system
The component-based system is too complex for me to work around.

For every page you add, if it is not related to other pages, you add 4 cmpnt classes.

Think of the amount of classes (and spaghetti code) you have to track if you eventually include session and database management.

Server-side validation is also hard to do.